ref(*)!: Refactors loader to use the new common package
