+++ ci/run-ci.sh
@@ -2,6 +2,6 @@
 
 set -ex
 
-cargo clippy --workspace
-cargo test --workspace
+cargo clippy --workspace --all-features
+cargo test --workspace --all-features
 (cd crates/wasm-pkg-loader/tests/e2e && cargo run)
