+++ crates/wasm-pkg-common/Cargo.toml
@@ -7,9 +7,15 @@ version.workspace = true
 authors.workspace = true
 license.workspace = true
 
+[features]
+metadata-client = ["dep:reqwest"]
+
 [dependencies]
+anyhow = "1.0"
 dirs = "5.0.1"
 http = "1.1.0"
+reqwest = { version = "0.12.0", featuers = ["json"], optional = true }
+semver = "1.0.23"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 toml = "0.8.13"
