+++ crates/wasm-pkg-common/src/lib.rs
@@ -1,33 +1,52 @@
 use http::uri::InvalidUri;
+use label::Label;
 
 pub mod config;
-mod label;
-mod package;
+pub mod label;
+pub mod metadata;
+pub mod package;
 pub mod registry;
-
-use label::InvalidLabel;
-pub use registry::Registry;
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     #[error("error reading config file: {0}")]
     ConfigFileIoError(#[source] std::io::Error),
+    #[error("failed to get registry credentials: {0:#}")]
+    CredentialError(#[source] anyhow::Error),
     #[error("invalid config: {0}")]
-    InvalidConfig(#[source] Box<dyn std::error::Error>),
+    InvalidConfig(#[source] anyhow::Error),
+    #[error("invalid content: {0}")]
+    InvalidContent(String),
+    #[error("invalid content digest: {0}")]
+    InvalidContentDigest(String),
+    #[error("invalid package manifest: {0}")]
+    InvalidPackageManifest(String),
     #[error("invalid package pattern: {0}")]
     InvalidPackagePattern(String),
     #[error("invalid label: {0}")]
-    InvalidLabel(#[from] InvalidLabel),
+    InvalidLabel(#[from] label::InvalidLabel),
     #[error("invalid package ref: {0}")]
     InvalidPackageRef(String),
     #[error("invalid registry: {0}")]
     InvalidRegistry(#[from] InvalidUri),
     #[error("invalid registry metadata: {0}")]
-    InvalidRegistryMetadata(#[source] serde_json::Error),
+    InvalidRegistryMetadata(#[source] anyhow::Error),
+    #[error("invalid version: {0}")]
+    InvalidVersion(#[from] semver::Error),
+    #[error("IO error: {0}")]
+    IoError(#[from] std::io::Error),
+    #[error("no registry configured for namespace {0:?}")]
+    NoRegistryForNamespace(Label),
+    #[error("registry error: {0}")]
+    RegistryError(#[source] anyhow::Error),
+    #[error("registry metadata error: {0:#}")]
+    RegistryMetadataError(#[source] anyhow::Error),
+    #[error("version not found: {0}")]
+    VersionNotFound(semver::Version),
 }
 
 impl Error {
-    fn invalid_config(err: impl std::error::Error + 'static) -> Self {
-        Self::InvalidConfig(Box::new(err))
+    fn invalid_config(err: impl Into<anyhow::Error>) -> Self {
+        Self::InvalidConfig(err.into())
     }
 }
