+++ crates/wkg/Cargo.toml
@@ -14,5 +14,6 @@ tempfile = "3.10.1"
 tokio = { workspace = true, features = ["macros", "rt"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
+wasm-pkg-common = { workspace = true }
 wasm-pkg-loader = { workspace = true }
 wit-component = "0.207"
